# Challenge4/docs/openapi/appointments.yaml
openapi: 3.1.0
info:
  title: Vitalis Clinic — Appointments Service
  version: 1.0.0
  description: >
    Máquina de estados de la cita, disponibilidad de médicos y panel admin.
    Dueño exclusivo de `Appointment`, `AppointmentEvent`, `Service`,
    `DoctorAvailability` y `UserProjection` (caché local de usuarios,
    poblada vía el evento `UserRegistered`). Orquesta el pago llamando a
    Payments por HTTP interno — nunca llama a Stripe directo. Ver
    RFC-001-bounded-contexts.md y ADR-009/ADR-010.
servers:
  - url: http://localhost:4000
    description: Gateway local (Docker Compose)
  - url: http://localhost:3002
    description: Servicio directo, local (fuera del gateway)

tags:
  - name: appointments
  - name: doctors
  - name: admin

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error, statusCode]
      properties:
        error:
          type: string
        statusCode:
          type: integer
    AppointmentStatus:
      type: string
      enum: [pending, confirmed, paid, reminded, completed, cancelled, no_show]
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        holdExpiresAt:
          type: string
          format: date-time
          nullable: true
        stripeCheckoutSessionId:
          type: string
          nullable: true
          description: Referencia informativa — Payments es la fuente de verdad del pago.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAppointmentRequest:
      type: object
      required: [doctorId, serviceId, startsAt]
      properties:
        doctorId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
    CreateAppointmentResponse:
      type: object
      properties:
        appointment:
          $ref: '#/components/schemas/Appointment'
        checkoutUrl:
          type: string
          format: uri
          description: Sesión de Stripe Checkout, obtenida de Payments vía HTTP interno.
    AppointmentEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
        type:
          type: string
          example: payment_confirmed
        payload:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    Doctor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        durationMin:
          type: integer
        priceCents:
          type: integer
    DoctorsResponse:
      type: object
      properties:
        doctors:
          type: array
          items:
            $ref: '#/components/schemas/Doctor'
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
    Slot:
      type: object
      properties:
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Body/query/params inválidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Sin token o token inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Autenticado pero sin el rol requerido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: >
        Operación no realizable en el estado actual (ej. reembolso falló al
        cancelar — la cita no cambia de estado, ver ADR-010).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

security:
  - bearerAuth: []

paths:
  /appointments:
    post:
      tags: [appointments]
      summary: Reservar una cita (solo pacientes)
      description: >
        Crea la cita en `pending` y llama internamente a Payments para
        obtener una sesión de checkout de Stripe (ver ADR-009: única
        dependencia HTTP síncrona del sistema).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          description: Cita creada, con URL de checkout.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAppointmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Payments no respondió (ver ADR-009, postmortem-payments-outage.md).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /appointments/me:
    get:
      tags: [appointments]
      summary: Mis citas (paciente autenticado)
      responses:
        '200':
          description: Lista de citas del paciente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /appointments/{id}/cancel:
    post:
      tags: [appointments]
      summary: Cancelar una cita (dueño de la cita o staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cita cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /doctors:
    get:
      tags: [doctors]
      summary: Listar médicos y servicios disponibles
      responses:
        '200':
          description: Médicos y servicios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /doctors/{id}/availability:
    get:
      tags: [doctors]
      summary: Horarios disponibles de un médico en una fecha
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        '200':
          description: Slots disponibles.
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/Slot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /doctors/me/agenda:
    get:
      tags: [doctors]
      summary: Mi agenda (solo médicos)
      responses:
        '200':
          description: Citas del médico autenticado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /admin/appointments:
    get:
      tags: [admin]
      summary: Listar todas las citas (solo staff)
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - name: date
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        '200':
          description: Citas con datos de paciente/médico resueltos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /admin/appointments/{id}/events:
    get:
      tags: [admin]
      summary: Log de eventos de una cita (solo staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Eventos ordenados cronológicamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /admin/appointments/{id}/cancel:
    post:
      tags: [admin]
      summary: Cancelar una cita como staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cita cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'